requirements.txt
discord
youtube_dl
PyNaCl
python-dotenv
scrapy